+++ packages/web/hooks/index.ts
@@ -2,5 +2,5 @@ export * from "./use-deterministic";
 export * from "./use-keplr";
 export * from "./use-boolean-with-window-event";
 export * from "./use-ibc-transfer";
-export * from "./use-basic-amount-config";
+export * from "./use-amount-config";
 export * from "./use-fake-fee-config";
